feat: extract out components
