+++ src/utils/components/ResetPasswordPage.tsx
@@ -1,21 +1,8 @@
 import type { FC } from "hono/jsx";
-import { Context } from "hono";
-import { Var } from "../types/Var";
-import { Env } from "../types/Env";
-import Layout from "./components/Layout";
+import Layout from "./Layout";
+import Button from "./Button";
 
-const Button: FC<{}> = ({ children }) => {
-  return (
-    <button
-      class="relative w-full rounded-lg text-center px-4 py-5 bg-primary text-textOnPrimary hover:bg-primaryHover text-base sm:mt-4 md:text-base"
-      type="submit"
-    >
-      <span class="flex items-center justify-center space-x-2">{children}</span>
-    </button>
-  );
-};
-
-export const ResetPasswordPage: FC<{ error?: string }> = ({ error }) => {
+const ResetPasswordPage: FC<{ error?: string }> = ({ error }) => {
   return (
     <Layout title="Reset Password">
       <div class="mb-8 text-2xl font-medium">Reset password</div>
@@ -34,3 +21,5 @@ export const ResetPasswordPage: FC<{ error?: string }> = ({ error }) => {
     </Layout>
   );
 };
+
+export default ResetPasswordPage;
